Refactor types, exports and main function name
