+++ src/index.ts
@@ -1,10 +1,10 @@
-import {ConfigParams} from './base-types';
+import {ConfigParams, Shopify} from './base-types';
 import {validateConfig} from './config';
 
 export * from './error';
 export * from './types';
 
-export function Shopify(config: ConfigParams) {
+export function shopifyApi(config: ConfigParams): Shopify {
   const validatedConfig = validateConfig(config);
   return {
     config: validatedConfig,
